backend/services/atomic/schedule/requirements.txt
fastapi==0.115.12
supabase==2.15.2
python-dotenv==1.0.1
uvicorn==0.34.0